I made a Queue struct and added functions and doc.
I thought it would be handy to have a Queue struct to keep track of whole queues, especially their head and their tail since we'l be accessing both a lot for enqueuing and dequeuing, respectively. We could in theory just keep track of one or the other, but then we would have to iterate through the whole queue each time we needed the one we didn't track, so I figured it was more efficient this way.

I was unsure of whether or not it would be helpful to keep track of the size of the Queues, so I left that as a commented line in the Queue struct declaration.

I also changed the payload to a `void*` so that we can put anything we want in the queue. 

Finally, I added or renamed and documented all our functions. However, I was unsure of your intentions with `rotateQ`; what should that do and is there a reason it was named with a "Q" instead of "Queue" like the others?